Fix possible dangling pointer to shader source
Force the constructed shader source to be dropped only after the
call to gl::ShaderSource.

On Windows x86_64 systems, compiling the shader results in OpenGL
returning syntax errors. The character on which it errors changes
between runs, and includes undefined characters, indicating the
source is invalid. This is a result of the string being dropped
before being copied to the shader object.